pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openclaw-godot"
version = "0.1.0"
description = "Autonomous AI agent framework for Godot Engine"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Derrick Barra"}
]
keywords = ["godot", "ai", "agents", "automation", "openclaw"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
requires-python = ">=3.10"
dependencies = [
    "mss>=9.0.0",
    "Pillow>=10.0.0",
    "PyAutoGUI>=0.9.54",
    "websockets>=12.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/derrickbarra/openclaw-godot"
Repository = "https://github.com/derrickbarra/openclaw-godot"
Issues = "https://github.com/derrickbarra/openclaw-godot/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
